Include isolates in roots & leaves
The exclusion of isolates for roots/leaves occured in commit
be0742661c5ec7e95ca570c2583df3701a45143a in the private repository
where nxontology was originally developed (as part of PR 219).
The reason for this exclusion is not clear and might have been a
mistake. According to https://stackoverflow.com/a/36321517/4651668
isolates are both roots and leaves.

This change does affect intrinsic_ic_sanchez for graphs with
isolates (disconnected nodes).